refactor: deduplicate copy_dir_all into nestweaver-storage
Moves the recursive directory copy function to nestweaver_storage::copy_dir_all
as a public utility. Adds explicit symlink skipping with a tracing warning.
Removes duplicate definitions from snapshot.rs and local.rs.
