+++ crates/nestweaver-storage/src/local.rs
@@ -1,20 +1,6 @@
 use crate::backend::{SnapshotMeta, StorageBackend};
+use crate::copy_dir_all;
 use std::path::{Path, PathBuf};
-
-fn copy_dir_all(src: &Path, dst: &Path) -> Result<(), anyhow::Error> {
-    std::fs::create_dir_all(dst)?;
-    for entry in std::fs::read_dir(src)? {
-        let entry = entry?;
-        let ty = entry.file_type()?;
-        let target = dst.join(entry.file_name());
-        if ty.is_dir() {
-            copy_dir_all(&entry.path(), &target)?;
-        } else {
-            std::fs::copy(entry.path(), target)?;
-        }
-    }
-    Ok(())
-}
 
 pub struct LocalBackend {
     root: PathBuf,
